+++ src/Rows/RowsPlugin.cs
@@ -3,6 +3,7 @@ using BepInEx;
 using BepInEx.Configuration;
 using BepInEx.Logging;
 using HarmonyLib;
+using Jotunn.Managers;
 using Jotunn.Utils;
 
 namespace Rows
@@ -60,6 +61,8 @@ namespace Rows
                     new ConfigurationManagerAttributes { IsAdminOnly = true }));
 
             Localizations.Register();
+
+            CommandManager.Instance.AddConsoleCommand(new RowsCommand());
 
             _harmony = new Harmony(PluginGuid);
             _harmony.PatchAll(typeof(RowsPlugin).Assembly);